backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
firebase-admin==6.2.0
pandas==2.1.0
numpy==1.26.4
python-dotenv==1.0.0
pydantic==2.4.0
python-multipart==0.0.6
librosa==0.10.0
soundfile==0.12.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.103.0
requests==2.31.0
gdown==5.2.0
google-cloud-storage==2.18.2
